requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
motor==3.3.2
orjson==3.9.14
packaging==23.2
pluggy==1.4.0
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
pymongo==4.6.1
pytest==8.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
